# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenDesk API
  description: >
    Sovereign, open-source office suite API. Provides document management,
    real-time collaboration, sharing, audit trails, workflow automation,
    cross-sovereign federation, local AI assistance, and verifiable data erasure.
  version: 0.1.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:5500
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT or dev-mode token passed as Bearer header
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
        issues:
          type: array
          items:
            type: object
      required: [error, issues]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
      required: [ok]

    # --- Documents ---
...
    Folder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        parentId:
          type: string
          format: uuid
          nullable: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateFolderBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        parentId:
          type: string
          format: uuid
          nullable: true
    RenameFolderBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200

    # --- Templates ---
...
    Reference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          type: integer
          nullable: true
        source:
          type: string
          nullable: true
        type:
          type: string
          enum: [article, book, chapter, website, other]
        doi:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
        url:
          type: string
          format: uri
          nullable: true
        publisher:
          type: string
          nullable: true
...
security:
  - BearerAuth: []
paths:
...
  /api/folders:
    get:
      tags: [Folders]
      summary: List folders
      description: List folders at root or under a parent. Results filtered by user permissions.
      parameters:
        - name: parentId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Array of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
    post:
      tags: [Folders]
      summary: Create a folder
      description: Create a new folder. Auto-grants owner role to creator.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderBody"
      responses:
        "201":
          description: Created folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /api/folders/{id}:
    put:
      tags: [Folders]
      summary: Rename a folder
      description: Rename an existing folder. Requires write permission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameFolderBody"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Folders]
      summary: Delete a folder
      description: Delete a folder. Requires delete permission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ===== Templates =====
...
  /api/references/import:
    post:
      tags: [References]
      summary: Import references from BibTeX or RIS
      description: Parse and import references from BibTeX or RIS format. Set Content-Type to application/x-bibtex or application/x-ris.
      requestBody:
        required: true
        content:
          application/x-bibtex:
            schema:
              type: string
          application/x-ris:
            schema:
              type: string
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  references:
                    type: array
                    items:
                      $ref: "#/components/schemas/Reference"
        "403":
          description: No write access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: Unsupported format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: No valid references found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/references/export:
    get:
      tags: [References]
      summary: Export references as BibTeX or RIS
      description: Export the entire reference library in the specified format.
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [bibtex, ris]
      responses:
        "200":
          description: Exported file
          content:
            application/x-bibtex:
              schema:
                type: string
            application/x-ris:
              schema:
                type: string
        "400":
          description: Invalid format parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: No read access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ===== Files =====
...